flask/requirements.txt
uwsgi
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
facenet-pytorch==2.3.0
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.0
nbconvert==5.6.1
nbformat==5.0.4
numpy==1.19.1
opencv-python==4.3.0.36
parso==0.6.1
Pillow==7.0.0
Pygments==2.5.2
pyparsing==2.4.7
pypng==0.0.20
pyrsistent==0.15.7
python-dateutil==2.8.1
pytubeX==0.1.1
PyYAML==5.3.1
pyzmq==18.1.1
requests==2.23.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
terminado==0.8.3
testpath==0.4.4
torch==1.4.0
torchvision==0.5.0
typing-extensions==3.7.4.2
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==1.0.1
wincertstore==0.2
